feat(all-tracks): add getSongList to /#/all-tracks
I was able to add the getSongList feature to the AllTracks page compo-
nent, bringing life to the route. A really satisfying experience it was.
The next step now is to add audio to the page so that it can start pla-
ying.

PS: I still need to go back and learn asynchronous JavaScript.