feat: add ui for albums, genre and artist pages
I've created a user interface for the 'Albums, 'Genre', and 'Artists'
pages. This will allow me to visualize what is going on when it is time
to load the data from the backend. The design is 'pretty good' so far.
The 'Albums' page doesn't look good with a list design, while the
'Genre' and 'Artists' pages look way better with the grid design.

What's next for me now to improve the 'Albums' page design and create a UI
for the 'Playlist' page. After that, then I'll begin to send data to it
to see how it works and improve it to make it better.

See you in the next commit 👋